:recycle: refactor: adiciona metodo de pagamentos no arquivo order do root e importando na order do pagarme
